create AvatarBase
